# packages/server/openapi.yaml
openapi: 3.0.3
info:
  title: AgentTrace API
  version: 0.1.0
  description: >
    AgentTrace is an observability platform for AI agents. It ingests telemetry
    events (LLM calls, completions, heartbeats, errors, and custom events),
    tracks agent health, computes usage statistics, and manages alert rules with
    delivery via webhook or email.
servers:
  - url: http://localhost:3000
    description: Local development server
security:
  - bearerAuth: []
  - apiKeyAuth: []
...
paths:
...
  /api/events:
    post:
      operationId: ingestEvents
      tags: [Events]
      summary: Ingest events
      description: >
        Accepts a single event object or a batch of events via the `events`
        array wrapper. Each event is validated independently; partial success
        returns HTTP 207. Rate-limited.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/EventBatchRequest"
                - $ref: "#/components/schemas/AgentEventInput"
      responses:
        "200":
          description: All events ingested successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventIngestionResponse"
        "207":
          description: Partial success -- some events failed validation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventIngestionResponse"
        "400":
          description: All events failed validation or no events provided
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: "#/components/schemas/EventIngestionResponse"
                  - $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "429":
          description: Rate limit exceeded

    get:
      operationId: queryEvents
      tags: [Events]
      summary: Query events
      description: >
        Returns events for a given agent, with optional filtering by time range,
        event type, provider, model, trace ID, and free-text search.
      parameters:
        - name: agent_id
          in: query
          required: true
          description: The agent_id to query events for
          schema:
            type: string
        - name: from
          in: query
          required: false
          description: Start of time range (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          description: End of time range (ISO 8601)
          schema:
            type: string
            format: date-time
        - name: event_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EventType"
        - name: provider
          in: query
          required: false
          schema:
            type: string
        - name: model
          in: query
          required: false
          schema:
            type: string
        - name: trace_id
          in: query
          required: false
          schema:
            type: string
        - name: search
          in: query
          required: false
          description: Free-text search across model, provider, error_message, and tags
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of events to return (default 1000)
          schema:
            type: integer
            default: 1000
      responses:
        "200":
          description: Matching events
          content:
            application/json:
              schema:
                type: object
                required:
                  - events
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentEvent"
        "400":
          description: Missing required agent_id parameter
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                error: agent_id query parameter is required
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/events/export:
    get:
      operationId: exportEvents
      tags: [Events]
      summary: Export events as JSON or CSV
      description: >
        Exports events for the given agent in either JSON or CSV format.
        The response includes a Content-Disposition header for file download.
      parameters:
        - name: agent_id
          in: query
          required: true
          description: The agent_id to export events for
          schema:
            type: string
        - name: format
          in: query
          required: false
          description: Export format
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: from
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: event_type
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/EventType"
        - name: provider
          in: query
          required: false
          schema:
            type: string
        - name: model
          in: query
          required: false
          schema:
            type: string
        - name: trace_id
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Exported event data
          headers:
            Content-Disposition:
              schema:
                type: string
              description: 'Attachment filename, e.g. attachment; filename="events-<agentId>.csv"'
          content:
            application/json:
              schema:
                type: object
                required:
                  - events
                properties:
                  events:
                    type: array
                    items:
                      $ref: "#/components/schemas/AgentEvent"
            text/csv:
              schema:
                type: string
                description: CSV data with header row
        "400":
          description: Missing agent_id or invalid format
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ---------------------------------------------------------------------------
  # Stats
  # ---------------------------------------------------------------------------
...
components:
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: "Pass the API token as a Bearer token: `Authorization: Bearer <token>`"
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: "Pass the API token via the x-api-key header"

  # ---------------------------------------------------------------------------
  # Reusable Parameters
  # ---------------------------------------------------------------------------
...
  responses:
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            error: Unauthorized

  # ---------------------------------------------------------------------------
  # Schemas
  # ---------------------------------------------------------------------------
  schemas:
...
    Error:
      type: object
      required:
        - error
      properties:
        error:
          type: string
          description: Human-readable error message

    # -- Health --
...
    EventType:
      type: string
      enum: [llm_call, completion, heartbeat, error, custom]
      description: The type of agent event
    EventSource:
      type: string
      enum: [sdk, proxy]
      description: How the event was ingested
    AgentEventInput:
      type: object
      required:
        - agent_id
        - event_type
        - source
        - timestamp
      properties:
        agent_id:
          type: string
          description: Agent that produced this event
        event_type:
          $ref: "#/components/schemas/EventType"
        source:
          $ref: "#/components/schemas/EventSource"
        timestamp:
          type: string
          format: date-time
          description: When the event occurred (ISO 8601)
        provider:
          type: string
          nullable: true
          description: LLM provider name (e.g. openai, anthropic)
        model:
          type: string
          nullable: true
          description: Model identifier (e.g. gpt-4o, claude-3-opus)
        tokens_in:
          type: integer
          nullable: true
          description: Number of input/prompt tokens
        tokens_out:
          type: integer
          nullable: true
          description: Number of output/completion tokens
        tokens_total:
          type: integer
          nullable: true
          description: Total token count
        cost_usd:
          type: number
          format: double
          nullable: true
          description: Estimated cost in USD
        latency_ms:
          type: integer
          nullable: true
          description: Request latency in milliseconds
        status_code:
          type: integer
          nullable: true
          description: HTTP status code of the upstream call
        error_message:
          type: string
          nullable: true
          description: Error description if the call failed
        tags:
          type: object
          nullable: true
          additionalProperties: true
          description: Arbitrary key-value metadata
        trace_id:
          type: string
          nullable: true
          description: Distributed trace identifier
        span_id:
          type: string
          nullable: true
          description: Span identifier within a trace
        parent_span_id:
          type: string
          nullable: true
          description: Parent span identifier for nested spans
    AgentEvent:
      allOf:
        - $ref: "#/components/schemas/AgentEventInput"
        - type: object
          required:
            - id
            - created_at
          properties:
            id:
              type: string
              description: Server-generated unique event ID (UUID)
            created_at:
              type: string
              format: date-time
              description: When the event was persisted
    EventBatchRequest:
      type: object
      required:
        - events
      properties:
        events:
          type: array
          items:
            $ref: "#/components/schemas/AgentEventInput"
          description: Array of events to ingest
    EventIngestionResult:
      type: object
      required:
        - index
        - success
      properties:
        index:
          type: integer
          description: Original index of the event in the batch
        success:
          type: boolean
          description: Whether the event was ingested successfully
        id:
          type: string
          description: Assigned event ID (present when success is true)
        error:
          type: string
          description: Validation error message (present when success is false)
    EventIngestionResponse:
      type: object
      required:
        - status
        - event_ids
        - results
      properties:
        status:
          type: string
          enum: [ok, error]
          description: Overall ingestion status
        event_ids:
          type: array
          items:
            type: string
          description: IDs of successfully ingested events
        results:
          type: array
          items:
            $ref: "#/components/schemas/EventIngestionResult"
          description: Per-event results ordered by original index

    # -- Stats --
...